♻️  Clean up processing flow
